==> src/github-network-analysis/02_international-collaboration/03_country-networks/03_country-network-analysis.Rmd <==
---
title: "Country Network Analysis"
output: html_document
---

```{r loading edgelist data, warning=FALSE, echo=FALSE}
rm(list = ls())
# install.packages(c("tidyverse", "igraph", "visNetwork", "bc3net", 

# load packages 
for (pkg in c("tidyverse", "data.table", "R.utils", "RPostgreSQL", "igraph", "rworldmap",
              "cowplot", "maditr", "lubridate", "countrycode")) {library(pkg, character.only = TRUE)}

# connect to postgresql to get data (in rivanna)
conn <- dbConnect(drv = PostgreSQL(), 
                  dbname = "sdad", 
                  host = "10.250.124.195", 
                  port = 5432, 
                  user = Sys.getenv("db_userid"), 
                  password = Sys.getenv("db_pwd"))

# query the bipartite edgelist data from github data  
sna_intl_ctry_nodelist <- dbGetQuery(conn, "SELECT country, users, repos, commits, additions, deletions
                                            FROM gh.sna_intl_ctry_nodelist")

# query the bipartite edgelist data from github data  
sna_intl_ctry_edgelist <- dbGetQuery(conn, "SELECT country1, country2, ctrycode1, ctrycode2, year, repo_wts
                                            FROM gh.sna_intl_ctry_edgelist")

# disconnect from postgresql
dbDisconnect(conn)
```

```{r, fig.width=11}
# filter out the omitted_geographies 
omitted_geographies <- c("Multiple Countries", "Europe", "Asia", "The Americas", "Africa")
`%notin%` <- Negate(`%in%`)
sna_intl_ctry_nodelist <- sna_intl_ctry_nodelist %>% 
  filter(country %notin% omitted_geographies) 

sna_intl_ctry_edgelist_full <- sna_intl_ctry_edgelist %>% 
  select(country1, country2, repo_wts) %>% 
  mutate(weight = repo_wts) %>% select(-repo_wts)
  

# checking that our network comes together 
intl_ctry_network <- simplify(graph_from_data_frame(d=sna_intl_ctry_edgelist_full, 
                                                    vertices=sna_intl_ctry_nodelist, directed=FALSE)) 

# let's take a look at the full network 
plot(intl_ctry_network, 
     edge.color="#d3d3d3", 
     edge.curved=0.1,
     edge.label=FALSE,
     vertex.size=8,
     vertex.color="#E57200", 
     vertex.label.color="#232D4B",
     layout=layout_with_kk)

``` 

This plot shows what we would expect. It looks as if (almost?) every country has some type of collaboration. But, in classic igraph fashion, the graph is (almost?) unbearable to look at. Let's export the node/edgelist into Gephi first and then we conduct break the network down over time to see what the dynamics look like. Before we do that, however, we want to perform use the Newman modularity algorithm on this network since Gephi's default Louvain algorithm is stochastic and harder to replicate for some plotting we do later. 

```{r}

# add the community membership to the nodelist 
intl_ctry_nodelist_gephi <- sna_intl_ctry_nodelist
fstgrdy <- fastgreedy.community(intl_ctry_network)
intl_ctry_nodelist_gephi$fstgrdy_comm <- fstgrdy$membership
intl_ctry_nodelist_gephi <- intl_ctry_nodelist_gephi %>% 
  rename(region = country) 

#intl_ctry_nodelist_gephi %>% filter(fstgrdy_comm != 1 & fstgrdy_comm != 2)

intl_ctry_nodelist_gephi$country_code <- countrycode(intl_ctry_nodelist_gephi$region, 
            origin = 'country.name', 
            destination = 'iso2c') 

intl_ctry_nodelist_gephi <- intl_ctry_nodelist_gephi %>% 
  mutate(country_code = ifelse(test = str_detect(string = region, 
                               pattern = "Kosovo"), 
                               yes = "XK", no = country_code))

# recode user ranges into categories  
intl_ctry_nodelist_gephi$fstgrdy_comm <- cut(intl_ctry_nodelist_gephi$fstgrdy_comm, 
                                             breaks=c(0,1,2,100), 
                                             labels=c("1", "2", "3"))

# graph the communities on the world map to see if they need consolidating
map_data("world") %>% 
  mutate(country_code = countrycode(region, 
                                    origin = 'country.name', 
                                    destination = 'iso2c')) %>% 
  left_join(intl_ctry_nodelist_gephi,  
            by = "country_code") %>% 
  ggplot() +
  theme_minimal() +
  geom_polygon(aes(x = long, y = lat, 
                   fill = fstgrdy_comm, 
                   group = group), 
                   color = "black") + 
  coord_fixed(1.2) +
  guides(fill=FALSE) + 
  scale_fill_manual(values = c("#E57200", # orange 
                               "#628ed8", # light blue 
                               "#232D4B"  # dark blue  
                               )) +
  theme(plot.title = element_text(size = 13, hjust = 0.5), 
        axis.line=element_blank(),axis.text.x=element_blank(),
        axis.text.y=element_blank(),axis.ticks=element_blank(),
        axis.title.x=element_blank(), axis.title.y=element_blank()) +
  labs(title= " Modularity Groupings for Country-Country \n Collaboration Networks (GitHub, 2008-2018)")

```

```{r}

# we need to recode the country column to id for gephi 
intl_ctry_nodelist_gephi <- intl_ctry_nodelist_gephi %>% 
  rename(id = region) %>% 
  select(id, country_code, everything())

intl_ctry_nodelist_gephi

# we need to recode the country columns to source, target for gephi
intl_ctry_edgelist_gephi <- sna_intl_ctry_edgelist_full %>% 
  rename(source = country1, target = country2)

intl_ctry_edgelist_gephi

# export the full network nodelist and edgelist 
setwd("~/oss-data/intl-country-nets")
write_csv(intl_ctry_nodelist_gephi, "intl_ctry_nodelist_gephi.csv")
write_csv(intl_ctry_edgelist_gephi, "intl_ctry_edgelist_gephi.csv") 

```
# Creating Country-to-Country Network Visuals in Gephi
