==> src/github-network-analysis/02_international-collaboration/03_country-networks/07_intl-ctry-nets-cum-summary.Rmd <==
---
title: "International Collaboration Networks"
author: "Brandon Kramer, José Bayoán Santiago Calderón and Gizem Korkmaz"
output: html_document
---

```{r setup, include=FALSE}
rm(list = ls())
# install.packages(c("tidyverse", "igraph", "visNetwork", "bc3net", 
# "data.table", "R.utils", "RPostgreSQL", "cowplot", "maditr", "linkprediction", "poweRlaw"))

# load packages 
for (pkg in c("tidyverse", "igraph", "visNetwork", "data.table", "extrafont",
              "R.utils", "RPostgreSQL", "cowplot", "maditr", "lubridate", 
              "poweRlaw", "plotly","gridExtra", "rmarkdown", 
              "grid")) {library(pkg, character.only = TRUE)}

setwd("~/oss-data/intl-ctry-nets-cum")
global_netstats_cum <- read_rds("global_netstats_cum.rds")
full_nodelist_cum <- read_rds("full_nodelist_cum.rds")
kcore_data <- read_rds("kcore_data_cum.rds")
decomp_data <- read_rds("decomp_data_cum.rds")
```

```{r bar_graphs_by_cum, fig.width=9.5, fig.height=8, echo=FALSE}
network_nodes <- global_netstats_cum %>% 
  ggplot( aes(time, node_count), y=node_count) +
  geom_bar(stat="identity", fill = "#990000", width = 0.7) + 
  theme_bw() + theme(axis.title.x = element_blank())

network_edges <- global_netstats_cum %>% 
  ggplot( aes(time, edge_count), y=edge_count) +
  geom_bar(stat="identity", fill = "#DA3330", width = 0.7) + 
  theme_bw() + theme(axis.title.x = element_blank())

network_wtd_edges <- global_netstats_cum %>% 
  ggplot( aes(time, wtd_edge_count), y=wtd_edge_count) +
  geom_bar(stat="identity", fill = "#D96403", width = 0.7) +
  theme_bw() + theme(axis.title.x = element_blank())

network_kmax <- global_netstats_cum %>% 
  ggplot( aes(time, kmax), y=kmax) +
  geom_bar(stat="identity", fill = "#eaaa31", width = 0.7) + 
  theme_bw() + theme(axis.title.x = element_blank())

network_commits <- global_netstats_cum %>% 
  ggplot( aes(time, commits), y=commits) +
  geom_bar(stat="identity", fill = "#000058", width = 0.7) + 
  theme_bw() + theme(axis.title.x = element_blank())

network_isolates <- global_netstats_cum %>% 
  ggplot( aes(time, isolates), y=isolates) +
  geom_bar(stat="identity", fill = "#628ed8", width = 0.7) + 
  theme_bw() + theme(axis.title.x = element_blank())

network_triads <- global_netstats_cum %>% 
  ggplot( aes(time, triads_300), y=triads_300) +
  geom_bar(stat="identity", fill = "#1d7c6b", width = 0.7) + 
  theme_bw() + theme(axis.title.x = element_blank())

grid.arrange(network_nodes, network_commits, 
             network_edges,  network_isolates,  
             network_wtd_edges, network_triads, 
             network_kmax, nrow=4, ncol=2,
             top="Cumulative Growth in Node, Edge and Commit Counts for GitHub Networks")
```

```{r, fig.width=11.5, fig.height=3.75, echo=FALSE}
network_nodes <- global_netstats_cum %>% 
  ggplot( aes(time, node_count), y=node_count) +
  geom_bar(stat="identity", fill = "#000058", width = 0.6) + 
  theme_minimal() + 
  theme(axis.title.x = element_blank(),
        axis.text=element_text(size=12, family="Arial"),
        axis.title=element_text(size=16, family="Arial")) +
  labs(y = "Node Count") +
  scale_x_discrete(breaks = c(2008,2009,2010,2011,2012,2013,2014,2015,2016,2017,2018,2019), 
                     labels = c(2008,"","","","",2013,"","","","","",2019)) +
  scale_y_continuous(breaks = c(0,  50, 100, 150, 200, 250), 
                                labels = c(0, 50, 100, 150, 200, 250))

network_edges <- global_netstats_cum %>% 
  ggplot( aes(time, edge_count), y=edge_count) +
  geom_bar(stat="identity", fill = "#628ed8", width = 0.6) + 
  theme_minimal() + 
  theme(axis.title.x = element_blank(),
                          axis.text=element_text(size=12, family="Arial"),
                          axis.title=element_text(size=16, family="Arial")) + 
  labs(y = "Edge Count") +
  scale_x_discrete(breaks = c(2008,2009,2010,2011,2012,2013,2014,2015,2016,2017,2018,2019), 
                     labels = c(2008,"","","","",2013,"","","","","",2019)) +
  scale_y_continuous(breaks = c(0, 2500, 5000, 7500, 10000, 12500), 
                                labels = c(0, 2500, 5000, 7500, 10000, 12500))

network_commits <- global_netstats_cum %>% 
  ggplot( aes(time, commits), y=commits) +
  geom_bar(stat="identity", fill = "#D96403", width = 0.6) + 
  theme_minimal() + 
  theme(axis.title.x = element_blank(),
        axis.text=element_text(size=12, family="Arial"),
        axis.title=element_text(size=16, family="Arial")) +
  labs(y = "Commit Count (in Millions)") +
  scale_x_discrete(breaks = c(2008,2009,2010,2011,2012,2013,2014,2015,2016,2017,2018,2019), 
                     labels = c(2008,"","","","",2013,"","","","","",2019)) +
  scale_y_continuous(breaks = c(0, 100000000, 200000000, 300000000, 400000000, 500000000), 
                                labels = c(0, 100, 200, 300, 400, 500))

figure <- grid.arrange(network_nodes, network_edges, network_commits, 
             nrow=1, ncol=3, top=
             #textGrob("Figure X. Cumulative Growth in Node, Edge and Commit Counts for GitHub Country Networks (2008-2019)",
            #            gp=gpar(fontsize=16, family="Arial"))
             )
      
```

```{r global_networks_cum descriptive trends, fig.width=9.5, echo=FALSE, layout="l-body-outset"}

all_others<-ggplot(global_netstats_cum, aes(time)) + 
  geom_line(aes(y = node_count, colour = "Node Count", group= 1), size=2) + 
  geom_line(aes(y = edge_count/100, colour = "Edge Count (/100)", group= 2), size=2) +
  geom_line(aes(y = wtd_edge_count/1000, colour = "Wtd Edge Count (/1000)", group= 3), size=2) +
  #geom_line(aes(y = kmax, colour = "Kmax (*100)", group = 4), size=2) +
  geom_line(aes(y = commits/1000, colour = "Commits (/1000)", group= 5), size=2) +   
  geom_line(aes(y = isolates, colour = "Isolates", group = 6), size=2) +
  geom_line(aes(y = triads_300, colour = "Triads (/10000)", group = 7), size=2) + 
  scale_colour_manual(values = 
  c("#990000","#DA3330","#D96403",#"#eaaa31",
    "#000058","#628ed8","#1d7c6b")) +                   
  #c( "#990000", "#D96403", "#eaaa31", "#628ed8", "#1d7c6b", "#600080", "#DA3330")) + 
  labs(y = "Node Count", x = "Year", colour = "",
       title = "Cumulative Network Growth for GitHub Data (2008-2019)") +
  theme_minimal() + theme(axis.title.x = element_blank()) + 
  theme(legend.position = c(0.15,0.8))
all_others

descriptives <- global_netstats_cum %>% select(year, node_count, edge_count, 
                               wtd_edge_count, commits, isolates, triads_300,kmax); descriptives
paged_table(descriptives, options = list(rows.print = 15))
```

```{r}

119.7 / 22.4 # 5 fold increase 

277 / 22 # 13 fold increase 

114 / 35

12015 / 236

12029 / 1926

1987352 / 5391

```

```{r density cum, fig.width=9.5, fig.height=8, echo=FALSE, layout="l-body-outset"}
density_cum<-ggplot(global_netstats_cum) + 
  geom_line(aes(x=time, y=density,group=1), size=2) + 
  labs(y = "Density (Cumulative)", x = "Year", colour = "#990000") + 
  theme_minimal() + theme(axis.title.x = element_blank())
transitivity_cum<-ggplot(global_netstats_cum) + 
  geom_line(aes(x=time, y=transitivity, group=1), size=2) + 
  labs(y = "Transitivity (Cumulative)", x = "Year", colour = "#D96403") + 
  theme_minimal() + theme(axis.title.x = element_blank())
louvain_cum<-ggplot(global_netstats_cum) + 
  geom_line(aes(x=time, y=louvain, group=1), size=2) + 
  labs(y = "Modularity (Cumulative)", x = "Year", colour = "#628ed8") +
  theme_minimal() + theme(axis.title.x = element_blank())
louvain_cum_scaled<-ggplot(global_netstats_cum) + 
  geom_line(aes(x=time, y=louvain_scaled, group=1), size=2) + 
  labs(y = "Modularity (Scaled)", x = "Year", colour = "#1d7c6b") +
  theme_minimal() + theme(axis.title.x = element_blank())
louvain_cum_logged<-ggplot(global_netstats_cum) + 
  geom_line(aes(x=time, y=louvain_logged, group=1), size=2) + 
  labs(y = "Modularity (Logged)", x = "Year", colour = "#600080") + 
  theme_minimal() + theme(axis.title.x = element_blank())
dist_cum<-ggplot(global_netstats_cum) + 
  geom_line(aes(x=time, y=mean_distance, group=1), size=2) + 
  labs(y = "Mean Distance", x = "Year", colour = "#600080") + 
  theme_minimal() + theme(axis.title.x = element_blank())


all_others<-ggplot(global_netstats_cum, aes(time)) + 
  geom_line(aes(y = node_count, colour = "Node Count", group= 1), size=2) + 
  geom_line(aes(y = edge_count/100, colour = "Edge Count (/100)", group= 2), size=2) +
  geom_line(aes(y = wtd_edge_count/1000, colour = "Wtd Edge Count (/1000)", group= 3), size=2) +
  #geom_line(aes(y = kmax*100, colour = "Kmax (*100)", group = 4), size=2) +
  geom_line(aes(y = commits/1000, colour = "Commits (/1000)", group= 5), size=2) +   
  geom_line(aes(y = isolates, colour = "Isolates", group = 6), size=2) +
  geom_line(aes(y = triads_300/10000, colour = "Triads (/10000)", group = 7), size=2) +
  scale_colour_manual(values = 
  c("#990000","#DA3330","#D96403",#"#eaaa31",
    "#000058","#628ed8","#1d7c6b")) +                  
  #c( "#990000", "#D96403", "#eaaa31", "#628ed8", "#1d7c6b", "#600080", "#DA3330")) + 
  labs(y = "Node Count", x = "Year", colour = "",
       title = "Cumulative Network Growth") +
  theme_minimal() + theme(axis.title.x = element_blank()) + 
  theme(legend.position = c(0.25,0.65))

grid.arrange(density_cum, louvain_cum, 
             transitivity_cum, louvain_cum_scaled,  
             dist_cum, louvain_cum_logged, nrow=3, ncol=2,
             top="Network Measures")

network_descriptives <- global_netstats_cum %>% 
  select(year, density, transitivity,louvain, louvain_logged, louvain_scaled)
paged_table(network_descriptives, options = list(rows.print = 15))

```

```{r, fig.width=11, fig.height=2.2, echo=FALSE, layout="l-body-outset"}
density_cum<-ggplot(global_netstats_cum) + 
  geom_line(aes(x=time, y=density,group=1), size=1.5) + 
  labs(y = "Density", x = "Year", colour = "#990000") + 
  theme_minimal() + theme(axis.title.x = element_blank()) +
  scale_x_discrete(breaks = c(2008,2009,2010,2011,2012,2013,2014,2015,2016,2017,2018,2019), 
                     labels = c(2008,"","","","",2013,"","","","","",2019))
transitivity_cum<-ggplot(global_netstats_cum) + 
  geom_line(aes(x=time, y=transitivity, group=1), size=1.5) + 
  labs(y = "Transitivity", x = "Year", colour = "#D96403") + 
  theme_minimal() + theme(axis.title.x = element_blank()) +
  scale_x_discrete(breaks = c(2008,2009,2010,2011,2012,2013,2014,2015,2016,2017,2018,2019), 
                     labels = c(2008,"","","","",2013,"","","","","",2019))
fstgrdy_cum_logged<-ggplot(global_netstats_cum) + 
  geom_line(aes(x=time, y=fstgrdy_logged, group=1), size=1.5) + 
  labs(y = "Modularity (Logged)", x = "Year", colour = "#600080") + 
  theme_minimal() + theme(axis.title.x = element_blank()) +
  scale_x_discrete(breaks = c(2008,2009,2010,2011,2012,2013,2014,2015,2016,2017,2018,2019), 
                     labels = c(2008,"","","","",2013,"","","","","",2019))
dist_cum<-ggplot(global_netstats_cum) + 
  geom_line(aes(x=time, y=mean_distance, group=1), size=1.5) + 
  labs(y = "Mean Distance", x = "Year", colour = "#600080") + 
  theme_minimal() + theme(axis.title.x = element_blank()) +
  scale_x_discrete(breaks = c(2008,2009,2010,2011,2012,2013,2014,2015,2016,2017,2018,2019), 
                     labels = c(2008,"","","","",2013,"","","","","",2019))

grid.arrange(density_cum, dist_cum, transitivity_cum, fstgrdy_cum_logged, nrow=1, ncol=4)
```
```{r, fig.width=11, fig.height=2.2, echo=FALSE, layout="l-body-outset"}
mean_btw<-ggplot(global_netstats_cum) + 
  geom_line(aes(x=time, y=mean_btw,group=1), size=1.5) + 
  labs(y = "Density", x = "Year", colour = "#990000") + 
  theme_minimal() + theme(axis.title.x = element_blank()) +
  scale_x_discrete(breaks = c(2008,2009,2010,2011,2012,2013,2014,2015,2016,2017,2018,2019), 
                     labels = c(2008,"","","","",2013,"","","","","",2019))


centr_deg<-ggplot(global_netstats_cum) + 
  geom_line(aes(x=time, y=centr_deg, group=1), size=1.5) + 
  labs(y = "centr_deg", x = "Year", colour = "#D96403") + 
  theme_minimal() + theme(axis.title.x = element_blank()) +
  scale_x_discrete(breaks = c(2008,2009,2010,2011,2012,2013,2014,2015,2016,2017,2018,2019), 
                     labels = c(2008,"","","","",2013,"","","","","",2019))
centr_btw<-ggplot(global_netstats_cum) + 
  geom_line(aes(x=time, y=centr_btw, group=1), size=1.5) + 
  labs(y = "centr_btw", x = "Year", colour = "#D96403") + 
  theme_minimal() + theme(axis.title.x = element_blank()) +
  scale_x_discrete(breaks = c(2008,2009,2010,2011,2012,2013,2014,2015,2016,2017,2018,2019), 
                     labels = c(2008,"","","","",2013,"","","","","",2019))
centr_clo<-ggplot(global_netstats_cum) + 
  geom_line(aes(x=time, y=centr_clo, group=1), size=1.5) + 
  labs(y = "centr_clo", x = "Year", colour = "#600080") + 
  theme_minimal() + theme(axis.title.x = element_blank()) +
  scale_x_discrete(breaks = c(2008,2009,2010,2011,2012,2013,2014,2015,2016,2017,2018,2019), 
                     labels = c(2008,"","","","",2013,"","","","","",2019))
centr_eigen<-ggplot(global_netstats_cum) + 
  geom_line(aes(x=time, y=centr_eigen, group=1), size=1.5) + 
  labs(y = "centr_eigen", x = "Year", colour = "#600080") + 
  theme_minimal() + theme(axis.title.x = element_blank()) +
  scale_x_discrete(breaks = c(2008,2009,2010,2011,2012,2013,2014,2015,2016,2017,2018,2019), 
                     labels = c(2008,"","","","",2013,"","","","","",2019))

grid.arrange(centr_deg, centr_btw, centr_clo, centr_eigen, nrow=1, ncol=4)
```




```{r percentage-change, echo=FALSE, fig.width=9.5}

global_netstats_cum <- global_netstats_cum %>% 
  mutate(nodes_pctcng = (node_count/lag(node_count) - 1) * 100) %>% 
  mutate(edges_pctcng = (edge_count/lag(edge_count) - 1) * 100) %>% 
  mutate(wtd_edges_pctcng = (wtd_edge_count/lag(wtd_edge_count) - 1) * 100) %>% 
  mutate(kmax_pctcng = (kmax/lag(kmax) - 1) * 100) %>%
  mutate(commits_pctcng = (commits/lag(commits) - 1) * 100) %>% 
  mutate(isolates_pctcng = (isolates/lag(isolates) - 1) * 100) %>% 
  mutate(dyads_pctcng = (dyads/lag(dyads) - 1) * 100) %>% 
  select(time, year, node_count, nodes_pctcng, edge_count, edges_pctcng, 
         wtd_edge_count, wtd_edges_pctcng, commits, commits_pctcng, 
         isolates, isolates_pctcng, dyads, dyads_pctcng, kmax, kmax_pctcng)
#global_netstats_cum
global_netstats_cum %>% filter(time != 2008) %>% ggplot(aes(time)) + 
  geom_line(aes(y = nodes_pctcng, colour = "nodes_pctcng", group= 1), size=2) + 
  geom_line(aes(y = edges_pctcng, colour = "edges_pctcng", group= 2), size=2) +
  geom_line(aes(y = wtd_edges_pctcng, colour = "wtd_edges_pctcng", group= 3), size=2) +
  geom_line(aes(y = kmax_pctcng, colour = "kmax_pctcng", group= 4), size=2) +
  geom_line(aes(y = commits_pctcng, colour = "commits_pctcng", group= 5), size=2) +   
  geom_line(aes(y = isolates_pctcng, colour = "isolates_pctcng", group = 6), size=2) +
  geom_line(aes(y = dyads_pctcng, colour = "dyads_pctcng", group = 7), size=0.5) + 
  scale_colour_manual(values = 
  c("#990000","#DA3330","#D96403","#eaaa31","#628ed8","#1d7c6b","#600080")) + 
  labs(y = "Node Count", x = "Year", colour = "",
       title = "Percent Change in Descriptive Measures for GitHub Networks") + 
  theme_minimal() + theme(axis.title.x = element_blank()) + theme(legend.position = c(0.85,0.8))

percentage_change <- global_netstats_cum %>% 
  filter(year != 2008) %>% 
  select(year, nodes_pctcng, edges_pctcng, 
         wtd_edges_pctcng, commits_pctcng, 
         isolates_pctcng, dyads_pctcng, kmax_pctcng)
paged_table(percentage_change, options= list(rows.print = 15))
```

```{r k_core_graph, echo=FALSE, fig.width=9.5}
# graph all values 
kcore_graph <- kcore_data %>% 
  #drop_na() %>%
  ggplot(aes(core)) + 
  geom_line(aes(y = k08, colour = "k08", group= 1), size=0.5) + 
  geom_line(aes(y = k09, colour = "k09", group= 2), size=0.5) +
  geom_line(aes(y = k10, colour = "k10", group= 3), size=0.5) +
  geom_line(aes(y = k11, colour = "k11", group= 4), size=0.5) +
  geom_line(aes(y = k12, colour = "k12", group= 5), size=0.5) +
  geom_line(aes(y = k13, colour = "k13", group= 6), size=0.5) +
  geom_line(aes(y = k14, colour = "k14", group= 7), size=0.5) +
  geom_line(aes(y = k15, colour = "k15", group= 8), size=0.5) +
  geom_line(aes(y = k16, colour = "k16", group= 9), size=0.5) +
  geom_line(aes(y = k17, colour = "k17", group= 10), size=0.5) +
  geom_line(aes(y = k18, colour = "k18", group= 11), size=0.5) +
  labs(title = "K-Core Distributions for Cumulative GitHub Networks") + 
  theme_minimal() + theme(axis.title.x = element_blank()) + theme(legend.position = c(0.85,0.8)); ggplotly()
```
```{r k_core_graph_2, echo=FALSE, fig.width=9.5}
kcore_graph <- kcore_data %>% 
  drop_na() %>%
  ggplot(aes(core)) + 
  geom_line(aes(y = k08, colour = "k08", group= 1), size=0.5) + 
  geom_line(aes(y = k09, colour = "k09", group= 2), size=0.5) +
  geom_line(aes(y = k10, colour = "k10", group= 3), size=0.5) +
  geom_line(aes(y = k11, colour = "k11", group= 4), size=0.5) +
  geom_line(aes(y = k12, colour = "k12", group= 5), size=0.5) +
  geom_line(aes(y = k13, colour = "k13", group= 6), size=0.5) +
  geom_line(aes(y = k14, colour = "k14", group= 7), size=0.5) +
  geom_line(aes(y = k15, colour = "k15", group= 8), size=0.5) +
  geom_line(aes(y = k16, colour = "k16", group= 9), size=0.5) +
  geom_line(aes(y = k17, colour = "k17", group= 10), size=0.5) +
  geom_line(aes(y = k18, colour = "k18", group= 11), size=0.5) +
  labs(title = "K-Core Distributions for Cumulative GitHub Networks") + 
  theme_minimal() + theme(axis.title.x = element_blank()) + theme(legend.position = c(0.85,0.8)); ggplotly()
```

```{r decomp_graph, echo=FALSE, fig.width=9.5}
# graph all values 
decomp_graph <- decomp_data %>% 
  #drop_na() %>%
  ggplot(aes(order(decomp))) + 
  geom_point(aes(y = dc08, colour = "dc08", group= 1), size=0.5) + 
  geom_point(aes(y = dc09, colour = "dc09", group= 2), size=0.5) +
  geom_point(aes(y = dc10, colour = "dc10", group= 3), size=0.5) +
  geom_point(aes(y = dc11, colour = "dc11", group= 4), size=0.5) +
  geom_point(aes(y = dc12, colour = "dc12", group= 5), size=0.5) +
  geom_point(aes(y = dc13, colour = "dc13", group= 6), size=0.5) +
  geom_point(aes(y = dc14, colour = "dc14", group= 7), size=0.5) +
  geom_point(aes(y = dc15, colour = "dc15", group= 8), size=0.5) +
  geom_point(aes(y = dc16, colour = "dc16", group= 9), size=0.5) +
  geom_point(aes(y = dc17, colour = "dc17", group= 10), size=0.5) +
  geom_point(aes(y = dc18, colour = "dc18", group= 11), size=0.5) +
  labs(title = "Decomposition Tendencies for Cumulative GitHub Networks") + 
  theme_minimal() + theme(axis.title.x = element_blank()) + theme(legend.position = c(0.85,0.8)); ggplotly()
```

```{r decomp_graph_2, echo=FALSE, fig.width=9.5}
# graph all values 
decomp_graph <- decomp_data %>% 
  drop_na() %>%
  ggplot(aes(decomp)) + 
  geom_line(aes(y = dc08, colour = "dc08", group= 1), size=0.5) + 
  geom_line(aes(y = dc09, colour = "dc09", group= 2), size=0.5) +
  geom_line(aes(y = dc10, colour = "dc10", group= 3), size=0.5) +
  geom_line(aes(y = dc11, colour = "dc11", group= 4), size=0.5) +
  geom_line(aes(y = dc12, colour = "dc12", group= 5), size=0.5) +
  geom_line(aes(y = dc13, colour = "dc13", group= 6), size=0.5) +
  geom_line(aes(y = dc14, colour = "dc14", group= 7), size=0.5) +
  geom_line(aes(y = dc15, colour = "dc15", group= 8), size=0.5) +
  geom_line(aes(y = dc16, colour = "dc16", group= 9), size=0.5) +
  geom_line(aes(y = dc17, colour = "dc17", group= 10), size=0.5) +
  geom_line(aes(y = dc18, colour = "dc18", group= 11), size=0.5) +
  labs(title = "Decomposition Tendencies for Cumulative GitHub Networks") + 
  theme_minimal() + theme(axis.title.x = element_blank()) + theme(legend.position = c(0.85,0.8)); ggplotly()
```

```{r}
library(data.table)
kcore_data <- kcore_data[,-1]
test <- kcore_data
test[,12] <- rownames(test)
test <- test %>% rename(core = V12)
test$core <- paste0("deg", test$core, sep="") 
test <- test %>% select(core, everything())
test <- test %>% remove_rownames %>% column_to_rownames(var="core")
test <- as.data.frame(t(test))
test <- as_tibble(c("2008", "2009", "2010", "2011", "2012", 
            "2013", "2014", "2015", "2016", "2017", "2018")) %>% 
  bind_cols(test) %>% 
  rename(time = value)  

kcore_trends <- ggplot(test, mapping = aes(x = time)) + 
  geom_line(aes(y = deg1, colour = "deg1", group= 1), size=0.5) +
  geom_line(aes(y = deg2, colour = "deg2", group= 2), size=0.5) +
  geom_line(aes(y = deg3, colour = "deg3", group= 3), size=0.5) +
  geom_line(aes(y = deg4, colour = "deg4", group= 4), size=0.5) +
  geom_line(aes(y = deg5, colour = "deg5", group= 5), size=0.5) +
  geom_line(aes(y = deg6, colour = "deg6", group= 6), size=0.5) +
  geom_line(aes(y = deg7, colour = "deg7", group= 7), size=0.5) +
  geom_line(aes(y = deg8, colour = "deg8", group= 8), size=0.5) +
  geom_line(aes(y = deg9, colour = "deg9", group= 9), size=0.5) +
  geom_line(aes(y = deg10, colour = "deg10", group= 10), size=0.5) +
  geom_line(aes(y = deg11, colour = "deg11", group= 11), size=0.5) +
  geom_line(aes(y = deg12, colour = "deg12", group= 12), size=0.5) +
  geom_line(aes(y = deg13, colour = "deg13", group= 13), size=0.5) +
  geom_line(aes(y = deg14, colour = "deg14", group= 14), size=0.5) +
  geom_line(aes(y = deg15, colour = "deg15", group= 15), size=0.5); ggplotly()

```

```{r}
full_nodelist_cum %>% drop_na(deg08) %>% mutate(avg_deg = mean(deg08)) %>% select(avg_deg)
full_nodelist_cum %>% drop_na(deg0809) %>% mutate(avg_deg = mean(deg0809)) %>% select(avg_deg)
full_nodelist_cum %>% drop_na(deg0810) %>% mutate(avg_deg = mean(deg0810)) %>% select(avg_deg)
full_nodelist_cum %>% drop_na(deg0811) %>% mutate(avg_deg = mean(deg0811)) %>% select(avg_deg)
full_nodelist_cum %>% drop_na(deg0812) %>% mutate(avg_deg = mean(deg0812)) %>% select(avg_deg)
full_nodelist_cum %>% drop_na(deg0813) %>% mutate(avg_deg = mean(deg0813)) %>% select(avg_deg)
full_nodelist_cum %>% drop_na(deg0814) %>% mutate(avg_deg = mean(deg0814)) %>% select(avg_deg)
full_nodelist_cum %>% drop_na(deg0815) %>% mutate(avg_deg = mean(deg0815)) %>% select(avg_deg)
full_nodelist_cum %>% drop_na(deg0816) %>% mutate(avg_deg = mean(deg0816)) %>% select(avg_deg)
full_nodelist_cum %>% drop_na(deg0817) %>% mutate(avg_deg = mean(deg0817)) %>% select(avg_deg)
full_nodelist_cum %>% drop_na(deg0818) %>% mutate(avg_deg = mean(deg0818)) %>% select(avg_deg)





```


```{r}

ggplot(global_netstats_cum, aes(x=node_count, y=louvain)) +
    geom_point(shape=1) 

ggplot(global_netstats_cum, aes(x=log(node_count), y=louvain)) +
    geom_point(shape=1)

cor(global_netstats_cum$node_count, global_netstats_cum$louvain)

constant <- rep(0.5, 11)

growth <- c(0.1, 0.2, 0.3, 0.3, 0.4, 0.5, 0.6, 0.6, 0.7, 0.8, 0.9)
decline <- c(0.9, 0.8, 0.7, 0.7, 0.6, 0.5, 0.4, 0.3, 0.2, 0.2, 0.1)
growth_scaled <- growth / global_netstats_cum$node_count
growth_logged <- growth / log(global_netstats_cum$node_count)
decline_scaled <- decline / global_netstats_cum$node_count
decline_logged <- decline / log(global_netstats_cum$node_count)
time <- c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11)

test_data <- as.data.frame(cbind(growth, decline,g_vector, g_logged, d_vector, d_logged, time))

a <- ggplot(test_data, aes(x=time, y=growth_scaled)) + geom_line(shape=1)
b <- ggplot(test_data, aes(x=time, y=growth_logged)) + geom_line(shape=1)
c <- ggplot(test_data, aes(x=time, y=decline_scaled)) + geom_line(shape=1)
d <- ggplot(test_data, aes(x=time, y=decline_logged)) + geom_line(shape=1)

grid.arrange(a, b, c, d, nrow=2, ncol=2)

```



```{r}
setwd("~/oss-data/intl-ctry-nets-cum")
full_nodelist <- readRDS("full_nodelist_cum.rds") 

community_explore <- full_nodelist %>% 
  arrange(-deg_cent0819) %>% 
  select(country, starts_with("fstgrdy"))

add08 <- community_explore %>% group_by(fstgrdy_comm08) %>% count() %>% 
  rename_at(1,~"group") %>% drop_na() %>% mutate(year = "2008") %>% rename("value" = n)
add0809 <- community_explore %>% group_by(fstgrdy_comm0809) %>% count() %>% 
  rename_at(1,~"group") %>% drop_na() %>% mutate(year = "2009") %>% rename("value" = n)
add0810 <- community_explore %>% group_by(fstgrdy_comm0810) %>% count() %>% 
  rename_at(1,~"group") %>% drop_na() %>% mutate(year = "2010") %>% rename("value" = n)
add0811 <- community_explore %>% group_by(fstgrdy_comm0811) %>% count() %>% 
  rename_at(1,~"group") %>% drop_na() %>% mutate(year = "2011") %>% rename("value" = n)
add0812 <- community_explore %>% group_by(fstgrdy_comm0812) %>% count() %>% 
  rename_at(1,~"group") %>% drop_na() %>% mutate(year = "2012") %>% rename("value" = n)
add0813 <- community_explore %>% group_by(fstgrdy_comm0813) %>% count() %>% 
  rename_at(1,~"group") %>% drop_na() %>% mutate(year = "2013") %>% rename("value" = n)
add0814 <- community_explore %>% group_by(fstgrdy_comm0814) %>% count() %>% 
  rename_at(1,~"group") %>% drop_na() %>% mutate(year = "2014") %>% rename("value" = n)
add0815 <- community_explore %>% group_by(fstgrdy_comm0815) %>% count() %>% 
  rename_at(1,~"group") %>% drop_na() %>% mutate(year = "2015") %>% rename("value" = n)
add0816 <- community_explore %>% group_by(fstgrdy_comm0816) %>% count() %>% 
  rename_at(1,~"group") %>% drop_na() %>% mutate(year = "2016") %>% rename("value" = n)
add0817 <- community_explore %>% group_by(fstgrdy_comm0817) %>% count() %>% 
  rename_at(1,~"group") %>% drop_na() %>% mutate(year = "2017") %>% rename("value" = n)
add0818 <- community_explore %>% group_by(fstgrdy_comm0818) %>% count() %>% 
  rename_at(1,~"group") %>% drop_na() %>% mutate(year = "2018") %>% rename("value" = n)
add0819 <- community_explore %>% group_by(fstgrdy_comm0819) %>% count() %>% 
  rename_at(1,~"group") %>% drop_na() %>% mutate(year = "2019") %>% rename("value" = n)

new_data <- as_tibble(rbind(add08, add0809, add0810, add0811, add0812, add0813, add0814, add0815, add0816, add0817, add0818, add0819))
 
new_data$group <- recode(new_data$group, `1` = "A", `2` = "B", `3` = "C", `4` = "D", `5` = "E")
new_data$group <- as.factor(new_data$group)
new_data$year <- as.numeric(new_data$year)


test_data <- new_data %>% 
  filter(group == "A" | group == "B" | group == "C")




ggplot(test_data, aes(x=time, y=value, fill=group)) + 
    geom_area()
 
# create data
time <- as.numeric(rep(seq(1,7),each=7))  # x Axis
value <- runif(49, 10, 100)               # y Axis
group <- rep(LETTERS[1:7],times=7)        # group, one shape per group
data <- data.frame(time, value, group)


test_data
```
